Update Boost skill for Enums namespace move
