--- .github/workflows/stale-issues.yml
@@ -1,17 +1,0 @@
-name: "Close stale issues"
-on:
-    schedule:
-        - cron: "0 0 * * *"
-
-jobs:
-    stale:
-        runs-on: ubuntu-latest
-        steps:
-            - uses: actions/stale@v10.2.0
-              with:
-                  repo-token: ${{ secrets.GITHUB_TOKEN }}
-                  stale-issue-message: "This issue is stale because it has been open 21 days with no activity. Remove stale label or comment or this will be closed in 7 days"
-                  stale-issue-label: "stale"
-                  exempt-issue-labels: "bug,enhancement,help wanted,next release,revisit for next major version"
-                  days-before-stale: 21
-                  days-before-close: 7
