+++ src/Facades/Activity.php
@@ -13,8 +13,8 @@ use Spatie\Activitylog\PendingActivityLog;
  * @method static \Spatie\Activitylog\ActivityLogger by(\Illuminate\Database\Eloquent\Model|string|int|null $modelOrId)
  * @method static \Spatie\Activitylog\ActivityLogger causedByAnonymous()
  * @method static \Spatie\Activitylog\ActivityLogger byAnonymous()
- * @method static \Spatie\Activitylog\ActivityLogger event(string|\Spatie\Activitylog\ActivityEvent $event)
- * @method static \Spatie\Activitylog\ActivityLogger setEvent(string|\Spatie\Activitylog\ActivityEvent $event)
+ * @method static \Spatie\Activitylog\ActivityLogger event(string|\Spatie\Activitylog\Enums\ActivityEvent $event)
+ * @method static \Spatie\Activitylog\ActivityLogger setEvent(string|\Spatie\Activitylog\Enums\ActivityEvent $event)
  * @method static \Spatie\Activitylog\ActivityLogger withProperties(mixed $properties)
  * @method static \Spatie\Activitylog\ActivityLogger withProperty(string $key, mixed $value)
  * @method static \Spatie\Activitylog\ActivityLogger createdAt(\DateTimeInterface $dateTime)
